+++ src/main/java/kitchenpos/menugroup/infra/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.menugroup.infra;
 
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
