+++ src/test/java/kitchenpos/Fixtures.java
@@ -1,7 +1,7 @@
 package kitchenpos;
 
 import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
