+++ src/test/java/kitchenpos/menugroup/infra/InMemoryMenuGroupRepository.java
@@ -1,7 +1,4 @@
-package kitchenpos.application;
-
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+package kitchenpos.menugroup.infra;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -9,6 +6,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 
 public class InMemoryMenuGroupRepository implements MenuGroupRepository {
     private final Map<UUID, MenuGroup> menuGroups = new HashMap<>();
